chore: rename agent piece to ai piece
