+++ packages/pieces/community/ai/src/lib/actions/text/ask-ai.ts
@@ -6,9 +6,9 @@ import {
 } from '@activepieces/pieces-framework';
 import { ModelMessage, ToolSet, generateText, stepCountIs } from 'ai';
 import { spreadIfDefined } from '@activepieces/shared';
-import { aiProps } from '../common/props';
-import { AIProviderName } from '../common/types';
-import { anthropicSearchTool, openaiSearchTool, googleSearchTool, createAIModel } from '../common/ai-sdk';
+import { aiProps } from '../../common/props';
+import { AIProviderName } from '../../common/types';
+import { anthropicSearchTool, openaiSearchTool, googleSearchTool, createAIModel } from '../../common/ai-sdk';
 
 export const askAI = createAction({
   name: 'askAi',
